refactor(desktop): fold the ipc one-caller files into their handlers
